LayoutTests/imported/w3c/web-platform-tests/css/css-scrollbars/scrollbar-color-005-expected.txt

Test scrollbar-color: vertical scrollbar
themed

FAIL scrollbar-color yellow blue on body overrides ::-webkit-scrollbar on scrollable area assert_less_than: themed scrollWidth expected a number less than 200 but got 200
